# Conflicts:
#	iotdb-core/confignode/src/main/java/org/apache/iotdb/confignode/persistence/schema/CNSnapshotFileType.java
#	iotdb-core/confignode/src/main/java/org/apache/iotdb/confignode/persistence/schema/ConfignodeSnapshotParser.java

+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/auth/user/LocalFileUserAccessor.java
@@ -421,8 +421,12 @@ public class LocalFileUserAccessor implements IUserAccessor {
   @Override
   public void cleanUserFolder() {
     File[] files = SystemFileFactory.INSTANCE.getFile(userDirPath).listFiles();
-    for (File file : files) {
-      FileUtils.deleteFileIfExist(file);
+    if (files != null) {
+      for (File file : files) {
+        FileUtils.deleteFileIfExist(file);
+      }
+    } else {
+      LOGGER.warn("User folder not exists");
     }
   }
